Music/Transition_reccomender/dj_recommender/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.10.0
numpy==1.24.3
scipy==1.11.4
essentia==2.1.0
openl3==0.4.0
torch==2.0.1
soundfile==0.12.1
faiss-cpu==1.7.4
pydantic==2.5.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
jinja2==3.1.2
